psmq: merge topic and payload fields together
This is big change. Right now, psmq had 2 static buffers, one
for topic and one for payload. This somehow seemed like a good
idea, but it had a lot of drawbacks. One of the most was ammount
of data that was needed to be copied between queues. So if we had
TOPIC_LEN_MAX 30, PAYLOAD_MAX 30, and topic was "/t" so only 2
bytes, WHOLE 30 bytes of TOPIC_LEN_MAX was being sent.

Right now, TOPIC_LEN_MAX and PAYLOAD_MAX is merged into single
buffer and only necessary data is sent over queue. This also
gives user some more flexibility, as not he can send on single
queue both messages that has long topic but no data, and messages
on short topics with a lot of data.

It also simplified code.

Signed-off-by: Michał Łyszczek <michal.lyszczek@bofc.pl>