src: use utils.c from $top_srcdir directly via symlink
autotools does not really like when we are taking source files
(.c, not .h) from $top_srcdir, it has real trouble finding proper
files if not for normal compilation, then when using distcheck.

I couldn't find clean way to do it properly so I hacked it with
using symlink. Not best solution, but best workaround.

Signed-off-by: Michał Łyszczek <michal.lyszczek@bofc.pl>