all: make passing of queue names fully optional
psmq is targeted to work mainly (if not only) on embedded devices.
Creating multiple brokers on such systems is a very corner case.
Because of that let's make passing queue names fully optional.
psmq will use default names and generate names for clients. This
will simplify interfaces a little bit.

Signed-off-by: Michał Łyszczek <michal.lyszczek@bofc.pl>